+++ Lib/fontbakery/specifications/adobe_fonts.py
@@ -74,7 +74,7 @@ expected_check_ids = [
     'com.google.fonts/check/varfont/regular_ital_coord',
     'com.google.fonts/check/varfont/regular_opsz_coord',
     'com.google.fonts/check/varfont/bold_wght_coord',
-    'com.google.fonts/check/180',  # Does the number of glyphs in the loca table match the maxp table?
+    'com.google.fonts/check/loca/maxp_num_glyphs',
     'com.google.fonts/check/ttx-roundtrip',
     'com.google.fonts/check/fontbakery_version',
     'com.google.fonts/check/ftxvalidator_is_available',